refactor: agrego el componente Welcome para estandarizar
