+++ src/components/layout/Navbar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../assets/styles/components.scss';
+import '../../assets/styles/components.scss';
 
 export default function Navbar({ setOriginal }) {
     return (
