+++ src/components/layout/Sidebar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import '../assets/styles/components.scss';
+import '../../assets/styles/components.scss';
 
 export default function Sidebar({ materias, carrera }) {
     return (
