+++ src/pages/mate/Mate.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../../assets/styles/components.scss';
 import '../../assets/styles/mate.scss';
-import Sidebar from '../../components/Sidebar';
+import Sidebar from '../../components/layout/Sidebar';
 import Materia from '../../components/Materia';
 import planA from '../../assets/images/plan-aplicada-transp-3.png';
 import planP from '../../assets/images/plan-pura-transp.png';
